refactor(BlockPopUp) : move BlockPopUp to common/PopUp folder
